fix: vetorise batching works better
TODO: add tests
