+++ src/marqo/s2_inference/s2_inference.py
@@ -49,11 +49,11 @@ def vectorise(model_name: str, content: Union[str, List[str]], model_properties:
         if isinstance(content, str):
             vectorised = available_models[model_cache_key].encode(content, normalize=normalize_embeddings, **kwargs)
         else:
-            vectorised = []
+            vector_batches = []
             batch_size = read_env_vars_and_defaults(EnvVars.MARQO_MAX_VECTORISE_BATCH_SIZE)
             for batch in generate_batches(content, batch_size=batch_size):
-                vectorised.append(available_models[model_cache_key].encode(batch, normalize=normalize_embeddings, **kwargs))
-            functools.reduce(lambda x, y: x + y, generate_batches(content, batch_size=batch_size))
+                vector_batches.append(available_models[model_cache_key].encode(batch, normalize=normalize_embeddings, **kwargs))
+            vectorised = functools.reduce(lambda x, y: numpy.concatenate(x, y), vector_batches)
     except UnidentifiedImageError as e:
         raise VectoriseError(str(e)) from e
